+++ metadrive/component/pgblock/create_pg_block_utils.py
@@ -11,7 +11,7 @@ from metadrive.component.road_network import Road
 from metadrive.component.road_network.node_road_network import NodeRoadNetwork
 from metadrive.constants import LineType, LineColor, DrivableAreaProperty
 from metadrive.utils.math_utils import get_vertical_vector
-from metadrive.utils.scene_utils import check_lane_on_road
+from metadrive.utils.pg_utils.utils import check_lane_on_road
 
 
 def create_bend_straight(
