+++ metadrive/component/pgblock/fork.py
@@ -8,7 +8,7 @@ from metadrive.component.pgblock.pg_block import PGBlock
 from metadrive.component.pgblock.ramp import Ramp
 from metadrive.component.road_network import Road
 from metadrive.constants import Decoration, LineType
-from metadrive.utils.scene_utils import check_lane_on_road
+from metadrive.utils.pg_utils.utils import check_lane_on_road
 from metadrive.component.pg_space import ParameterSpace, Parameter, BlockParameterSpace
