+++ metadrive/component/road_network/node_road_network.py
@@ -7,7 +7,7 @@ from metadrive.component.road_network.base_road_network import BaseRoadNetwork
 from metadrive.component.road_network.road import Road
 from metadrive.constants import Decoration
 from metadrive.utils.math_utils import get_boxes_bounding_box
-from metadrive.utils.scene_utils import get_lanes_bounding_box
+from metadrive.utils.pg_utils.utils import get_lanes_bounding_box
 
 logger = logging.getLogger(__name__)
