+++ metadrive/component/vehicle_navigation_module/node_network_navigation.py
@@ -9,7 +9,7 @@ from metadrive.component.road_network import Road
 from metadrive.component.road_network.node_road_network import NodeRoadNetwork
 from metadrive.component.vehicle_navigation_module.base_navigation import BaseNavigation
 from metadrive.utils import clip, norm, get_np_random
-from metadrive.utils.scene_utils import ray_localization
+from metadrive.utils.pg_utils.utils import ray_localization
 from metadrive.component.pg_space import Parameter, BlockParameterSpace
