+++ metadrive/obs/top_down_renderer.py
@@ -11,7 +11,7 @@ from metadrive.utils.waymo_utils.waymo_type import WaymoLaneProperty
 from metadrive.engine.engine_utils import get_engine
 from metadrive.obs.top_down_obs_impl import WorldSurface, VehicleGraphics, LaneGraphics
 from metadrive.utils.interpolating_line import InterpolatingLine
-from metadrive.utils.map_utils import is_map_related_instance
+from metadrive.utils.utils import is_map_related_instance
 from metadrive.utils.utils import import_pygame
 from metadrive.utils.waymo_utils.utils import convert_polyline_to_metadrive
