+++ metadrive/tests/test_functionality/test_export_map.py
@@ -4,7 +4,7 @@ try:
     from metadrive.envs.real_data_envs.nuplan_env import NuPlanEnv
 except ImportError:
     pass
-from metadrive.utils.export_utils.utils import draw_map
+from metadrive.utils.scene_utils.utils import draw_map
 from metadrive.envs.metadrive_env import MetaDriveEnv
 from metadrive.envs.real_data_envs.waymo_env import WaymoEnv
 from metadrive.policy.idm_policy import WaymoIDMPolicy
