+++ metadrive/utils/pg_utils/generate_maps.py
