+++ metadrive/utils/scene_utils/type.py
