+++ metadrive/utils/utils.py
@@ -171,3 +171,15 @@ def get_object_from_node(node: BulletBodyNode):
         return get_object(ret)[ret]
     else:
         return ret
+
+
+def is_map_related_instance(obj):
+    from metadrive.component.block.base_block import BaseBlock
+    from metadrive.component.map.base_map import BaseMap
+    return True if isinstance(obj, BaseBlock) or isinstance(obj, BaseMap) else False
+
+
+def is_map_related_class(object_class):
+    from metadrive.component.block.base_block import BaseBlock
+    from metadrive.component.map.base_map import BaseMap
+    return True if issubclass(object_class, BaseBlock) or issubclass(object_class, BaseMap) else False
